Add support for 64-bit
[git-p4: depot-paths = "//depot/game/": change = 300]

+++ code/data/memcard/memorycardmanager.cpp
@@ -326,7 +326,7 @@ MemoryCardManager::Init( IRadDriveErrorCallback* radDriveErrorCallback )
 
         radDriveOpenAsync( &m_pDrives[ i ], SAVE_GAME_DRIVE[ i ] );
         rAssert( m_pDrives[ i ] );
-        m_pDrives[ i ]->AddCompletionCallback( this, reinterpret_cast<void*>( i ) );
+        m_pDrives[ i ]->AddCompletionCallback( this, reinterpret_cast<void*>( (uintptr_t)i ) );
 
         // register error handler
         //
@@ -868,7 +868,7 @@ MemoryCardManager::OnDriveOperationsComplete( void* pUserData )
 	}
     else if( m_currentState == STATE_OPENING_DRIVES )
     {
-        unsigned int driveIndex = reinterpret_cast<unsigned int>( pUserData );
+        unsigned int driveIndex = reinterpret_cast<uintptr_t>( pUserData );
         rAssert( driveIndex < NUM_SAVE_GAME_DRIVES );
 
         m_numDrivesOpened++;
