+++ libs/radscript/src/typeinfo/TypeInfoLoader.cpp
@@ -176,7 +176,7 @@ struct radTypeInfoLoader
 
         ref< IRadTypeInfo > xIRadTypeInfo;
 
-		switch( (FileLoadState) (int) pUserData )
+		switch( (FileLoadState) (intptr_t) pUserData )
 		{
 			case GetSize:
 			{
